+++ api/src/build.rs
@@ -3,11 +3,9 @@ use std::path::{Path, PathBuf};
 use std::process::Command;
 
 use anyhow::{anyhow, Context, Result};
-use cargo::core::compiler::CompileMode;
-use cargo::core::Workspace;
-use cargo::ops::CompileOptions;
 use rocket::tokio;
 use rocket::tokio::io::AsyncWriteExt;
+use shuttle_service::loader::build_crate;
 use uuid::Uuid;
 
 #[cfg(debug_assertions)]
@@ -189,32 +187,4 @@ fn extract_tarball(crate_path: &Path, project_path: &Path) -> Result<()> {
     } else {
         Ok(())
     }
-}
-
-/// Given a project directory path, builds the crate
-fn build_crate(project_path: &Path, buf: Box<dyn std::io::Write>) -> Result<PathBuf> {
-    let mut shell = cargo::core::Shell::from_write(buf);
-    shell.set_verbosity(cargo::core::Verbosity::Normal);
-
-    let cwd = std::env::current_dir()
-        .with_context(|| "couldn't get the current directory of the process")?;
-    let homedir = cargo::util::homedir(&cwd).ok_or_else(|| {
-        anyhow!(
-            "Cargo couldn't find your home directory. \
-                 This probably means that $HOME was not set."
-        )
-    })?;
-
-    let config = cargo::Config::new(shell, cwd, homedir);
-    let manifest_path = project_path.join("Cargo.toml");
-
-    let ws = Workspace::new(&manifest_path, &config)?;
-    let opts = CompileOptions::new(&config, CompileMode::Build)?;
-    let compilation = cargo::ops::compile(&ws, &opts)?;
-
-    if compilation.cdylibs.is_empty() {
-        return Err(anyhow!("a cdylib was not created"));
-    }
-
-    Ok(compilation.cdylibs[0].path.clone())
 }
